# dolci-reva-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dolcireva API",
        "description": "API de réservation et gestion d'établissements (Hôtels, Résidences, Restaurants, Lounges, Night Clubs)",
        "contact": {
            "name": "Support Dolcireva",
            "email": "support@dolcireva.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": "https://api.dolcireva.com/api",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/residences": {
            "get": {
                "tags": [
                    "Residences"
                ],
                "summary": "Liste des résidences",
                "description": "Récupère la liste de toutes les résidences de l'utilisateur authentifié",
                "operationId": "getResidences",
                "responses": {
                    "200": {
                        "description": "Liste des résidences récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Residence"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Residences"
                ],
                "summary": "Créer une résidence",
                "description": "Crée une nouvelle résidence",
                "operationId": "createResidence",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "address",
                                    "city",
                                    "country",
                                    "property_type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Villa de Luxe"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Magnifique villa avec piscine privée"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Rue de la Paix"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Nice"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "France"
                                    },
                                    "property_type": {
                                        "type": "string",
                                        "enum": [
                                            "APARTMENT",
                                            "HOUSE",
                                            "VILLA",
                                            "STUDIO"
                                        ],
                                        "example": "VILLA"
                                    },
                                    "rental_type": {
                                        "type": "string",
                                        "enum": [
                                            "DAILY",
                                            "WEEKLY",
                                            "MONTHLY"
                                        ],
                                        "example": "DAILY"
                                    },
                                    "bedrooms": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "bathrooms": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "max_guests": {
                                        "type": "integer",
                                        "example": 6
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 200
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 43.7102
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 7.262
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "amenities": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Résidence créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Residence created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Residence"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Residence": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "owner_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Résidence Moderne"
                    },
                    "description": {
                        "type": "string",
                        "example": "Appartement moderne en centre-ville"
                    },
                    "address": {
                        "type": "string",
                        "example": "654 Rue de la République"
                    },
                    "city": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "country": {
                        "type": "string",
                        "example": "France"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 48.8566
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 2.3522
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "STUDIO",
                            "T1",
                            "T2",
                            "T3",
                            "T4",
                            "T5+"
                        ],
                        "example": "T2"
                    },
                    "max_guests": {
                        "type": "integer",
                        "example": 4
                    },
                    "bedrooms": {
                        "type": "integer",
                        "example": 2
                    },
                    "bathrooms": {
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 80
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Une erreur s'est produite"
                    },
                    "error": {
                        "type": "string",
                        "example": "Error details"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "The field name is required."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification par token Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}